requirements.txt
sumo-rl==1.4.5
ray[rllib]==2.31.0
torch>=2.0.0
gymnasium==0.28.1
pettingzoo==1.24.3
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
tensorboard>=2.13.0
pyyaml>=6.0
tqdm>=4.65.0

